pyproject.toml
[project]
name = "hand-axes-image-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "piexif>=1.1.3",
    "pyyaml>=6.0.3",
]
